ajout des confirmations et un peut de stylle

+++ resources/views/carte_compte/list.blade.php
@@ -14,37 +14,43 @@
                         @csrf
                         <input type="hidden" name="carte_id" id="carte_id">
                         <div class="note">
-                          <p id="formtitle"><strong>Informations sur la carte</strong></p>
+                            <p id="formtitle"><strong>Informations sur la carte</strong></p>
                         </div>
                         <div class="row tout">
                             <div class="col right">
-                              <input type="number" name="numero_carte" id="numero_carte" class="form-control first" placeholder="numero de carte" aria-label="numero de carte">
-                              <select class="form-select first" name="type" aria-label="Default select example">
-                                <option selected>Type de carte</option>
-                                <option value="carte electron">Carte electron</option>
-                                <option value="carte visa">Carte visa</option>
-                                <option value="master carte">Master carte</option>
-                              </select>
-                              <input type="date" name="date_creation" id="date_creation" class="form-control first" placeholder="Select Date" aria-label="date_creation">
+                                <input type="number" name="numero_carte" id="numero_carte" class="form-control first"
+                                    placeholder="numero de carte" aria-label="numero de carte">
+                                <select class="form-select first" name="type" aria-label="Default select example">
+                                    <option selected>Type de carte</option>
+                                    <option value="carte electron">Carte electron</option>
+                                    <option value="carte visa">Carte visa</option>
+                                    <option value="master carte">Master carte</option>
+                                </select>
+                                <input type="date" name="date_creation" id="date_creation" class="form-control first"
+                                    placeholder="Select Date" aria-label="date_creation">
                             </div>
                             <div class="col gauche">
-                                <input type='number' name="codesecret" id="codesecret" class="form-control first" placeholder="code secret" aria-label="codesecret"/>
-                                <input type="date" name="date_expiration" id="date_expiration" class="form-control first" placeholder="Select Date" aria-label="date_expiration">
-                              <div>
+                                <input type='number' name="codesecret" id="codesecret" class="form-control first"
+                                    placeholder="code secret" aria-label="codesecret" />
+                                <input type="date" name="date_expiration" id="date_expiration" class="form-control first"
+                                    placeholder="Select Date" aria-label="date_expiration">
+                                <div>
                                     <p style="padding: 0;margin: 0;">Status</p>
-                                  <div class="form-check">
-                                    <input class="form-check-input sex" type="radio" name="statut" id="actif" value="1" checked>
-                                    <label class="form-check-label" for="actif">
-                                      Actif
-                                    </label>
-                                  </div>
-                                  <div class="form-check">
-                                    <input class="form-check-input sex" type="radio" name="statut" id="inactif" value="0" >
-                                    <label class="form-check-label" for="inactif">
-                                      Inactif
-                                    </label>
-                                  </div>
-                              </div>
+                                    <div class="form-check">
+                                        <input class="form-check-input sex" type="radio" name="statut" id="actif"
+                                            value="1" checked>
+                                        <label class="form-check-label" for="actif">
+                                            Actif
+                                        </label>
+                                    </div>
+                                    <div class="form-check">
+                                        <input class="form-check-input sex" type="radio" name="statut" id="inactif"
+                                            value="0">
+                                        <label class="form-check-label" for="inactif">
+                                            Inactif
+                                        </label>
+                                    </div>
+                                </div>
                             </div>
                         </div>
                     </form>
@@ -101,11 +107,18 @@
                 "bInfo": true,
                 "bFilter": true,
                 "bAutoWidth": false,
-                "aoColumns" : [
-                    { sWidth: '50px' },
-                    { sWidth: '100px' },
-                    { sWidth: '120px' },
-                    { sWidth: '30px' }
+                "aoColumns": [{
+                        sWidth: '50px'
+                    },
+                    {
+                        sWidth: '100px'
+                    },
+                    {
+                        sWidth: '120px'
+                    },
+                    {
+                        sWidth: '30px'
+                    }
                 ],
                 columns: [{
                         data: 'id',
@@ -170,8 +183,8 @@
             $('body').on('click', '#edit', function() {
                 var formdata = new FormData(form);
                 $.ajax({
-                    url:'{{route("carte.edit")}}',
-                    method:'POST',
+                    url: '{{ route('carte.edit') }}',
+                    method: 'POST',
                     data: formdata,
                     processData: false,
                     contentType: false,
@@ -188,32 +201,48 @@
             });
 
             // DELETE
-            $('body').on('click','#delete',function () {
-               var id = $(this).data('id');
-               $.ajax({
-                    url:'{{url("carte-delete")}}'+'/'+id,
-                    method:'DELETE',
-
-                    success:function (response) {
-                        table.ajax.reload();
-                        console.log(response);
-                    },
-                    error:function (error) {
-                        console.log(error);
-                    }
-                });
+            $('body').on('click', '#delete', function() {
+                var id = $(this).data('id');
+                swal({
+                        title: "Are you sure?",
+                        text: "Once deleted, you will not be able to recover this imaginary file!",
+                        icon: "warning",
+                        buttons: true,
+                        dangerMode: true,
+                    })
+                    .then((willDelete) => {
+                        if (willDelete) {
+                            $.ajax({
+                                url: '{{ url('carte-delete') }}' + '/' + id,
+                                method: 'DELETE',
+            
+                                success: function(response) {
+                                    swal("Poof! Your imaginary file has been deleted!", {
+                                        icon: "success",
+                                    });
+                                    table.ajax.reload();
+                                    console.log(response);
+                                },
+                                error: function(error) {
+                                    console.log(error);
+                                }
+                            });
+                        } else {
+                            swal("Your imaginary file is safe!");
+                        }
+                    });
             });
             // show carte
-            $('body').on('click','#detail',function() {
+            $('body').on('click', '#detail', function() {
                 var id = $(this).data('id');
                 $.ajax({
-                    url:'{{url("carte-toshow")}}'+'/'+id,
-                    method:'GET',
+                    url: '{{ url('carte-toshow') }}' + '/' + id,
+                    method: 'GET',
 
-                    success:function (response) {
-                        window.location.href = "{{url('carte-show')}}"+"/"+id;
+                    success: function(response) {
+                        window.location.href = "{{ url('carte-show') }}" + "/" + id;
                     },
-                    error:function(error) {
+                    error: function(error) {
                         console.log(error);
                     }
                 });
